Organise classes into task, command and duke packages
