+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import duke.Storage;
+import duke.Ui;
+import task.TaskList;
+
 public class ExitCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         ui.showGoodbye();
