+++ src/main/java/command/MarkCommand.java
@@ -1,18 +1,24 @@
-public class UnmarkCommand extends Command {
-    protected int index;
-    public static final String COMMAND_WORD = "unmark";
-    public static final String MESSAGE_SUCCESS = " OK, I've marked this task as not done yet:\n";
-    protected boolean isExit;
+package command;
 
-    public UnmarkCommand(int index) {
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import task.Task;
+import task.TaskList;
+
+public class MarkCommand extends Command {
+    protected int index;
+    public static final String COMMAND_WORD = "mark";
+    public static final String MESSAGE_SUCCESS = " Nice! I've marked this task as done:\n";
+
+    public MarkCommand(int index) {
         this.index = index;
-        this.isExit = false;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         if (this.index >= 0 && this.index < tasks.getSize()) {
             Task task = tasks.getTask(this.index);
-            tasks.unmarkTask(task);
+            tasks.markTask(task);
             storage.writeToFile(tasks.getList());
             ui.showMessage(MESSAGE_SUCCESS + "     " + task);
         } else { // user input is an integer bigger than size of task list
