+++ src/main/java/command/UnmarkCommand.java
@@ -1,16 +1,26 @@
-public class MarkCommand extends Command {
-    protected int index;
-    public static final String COMMAND_WORD = "mark";
-    public static final String MESSAGE_SUCCESS = " Nice! I've marked this task as done:\n";
+package command;
 
-    public MarkCommand(int index) {
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import task.Task;
+import task.TaskList;
+
+public class UnmarkCommand extends Command {
+    protected int index;
+    public static final String COMMAND_WORD = "unmark";
+    public static final String MESSAGE_SUCCESS = " OK, I've marked this task as not done yet:\n";
+    protected boolean isExit;
+
+    public UnmarkCommand(int index) {
         this.index = index;
+        this.isExit = false;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         if (this.index >= 0 && this.index < tasks.getSize()) {
             Task task = tasks.getTask(this.index);
-            tasks.markTask(task);
+            tasks.unmarkTask(task);
             storage.writeToFile(tasks.getList());
             ui.showMessage(MESSAGE_SUCCESS + "     " + task);
         } else { // user input is an integer bigger than size of task list
