+++ src/main/java/duke/Duke.java
@@ -1,3 +1,8 @@
+package duke;
+
+import command.Command;
+import task.TaskList;
+
 import java.time.format.DateTimeParseException;
 
 public class Duke {
