+++ src/main/java/duke/Parser.java
@@ -1,3 +1,7 @@
+package duke;
+
+import command.*;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
